Make the PreInterpreter work
The PreInterpreter is a small step to make the Interpreting faster because it already builds a pre-structure to make the Component Creations Functions easier to work.

There is still parts missing like: Includes and StringLiterals.

Includes are get handled by the PreInterpreter to include the full tree at the end.